examples/python/medical_story_requirements.txt
# Medical Imaging Story Requirements
# Core dependencies for the medical imaging story

# Google ADK
google-adk

# Medical imaging libraries
nibabel>=5.0.0
pydicom>=2.3.0
dicom2nifti>=2.4.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
Pillow>=8.3.0

# Data processing
pandas>=1.3.0
scikit-learn>=1.0.0

# Optional: For advanced medical imaging
# monai>=1.0.0
# SimpleITK>=2.1.0
